moved mod plugin background code to subsystems
