+++ coreplugins/mod.py
@@ -6,198 +6,9 @@ from discord.ext import commands
 from conf import Config
 from botutils import utils, permChecks, enums
 from Geckarbot import BasePlugin
-
-
-class Blacklist:
-    """Manage the user banlist for using the bot"""
-
-    def __init__(self, plugin):
-        self.plugin = plugin
-
-    def bl_conf(self):
-        return Config().get(self.plugin)['blacklist']
-
-    def add_user(self, user: discord.Member):
-        """Adds user to bot blacklist, returns True if added"""
-        if not self.is_member_on_blacklist(user):
-            self.bl_conf().append(user.id)
-            Config().save(self.plugin)
-            return True
-        else:
-            return False
-
-    def del_user(self, user: discord.Member):
-        """Deletes user to bot blacklist, returns True if deleted"""
-        if self.is_member_on_blacklist(user):
-            self.bl_conf().remove(user.id)
-            Config().save(self.plugin)
-            return True
-        else:
-            return False
-
-    def get_blacklist_names(self):
-        """Returns the blacklisted member names"""
-        blacklisted_members = ", ".join([self.plugin.bot.get_user(uid).name for uid in self.bl_conf()])
-        return blacklisted_members
-
-    def is_member_on_blacklist(self, user: discord.Member):
-        """Returns if user is on bot blacklist"""
-        return self.is_userid_on_blacklist(user.id)
-
-    def is_userid_on_blacklist(self, userid: int):
-        """Returns if user id is on bot blacklist"""
-        if userid in self.bl_conf():
-            return True
-        else:
-            return False
-
-
-class Greylist:
-    """Manage the user greylist for using the bot.
-    Users on greylist can't play the bot provided games on their greylist.
-    """
-
-    def __init__(self, plugin):
-        self.plugin = plugin
-
-    def gl_conf(self):
-        return Config().get(self.plugin)['greylist']
-
-    def add(self, user: discord.Member, game: enums.GreylistGames = None):
-        """Adds the given game to users greylist.
-        If game is None, all games will be added.
-        If user is already on list, the user game list will be updated.
-        If user is new on list, True will be returned, otherwise False.
-        """
-        if game is None:
-            game = enums.GreylistGames.ALL
-        was_added = True
-        if self.is_user_on_greylist(user, game):
-            game = self.gl_conf()[user.id] | game
-            was_added = False
-        self.gl_conf()[user.id] = game
-        Config().save(self.plugin)
-        return was_added
-
-    def remove(self, user: discord.Member, game: enums.GreylistGames = None):
-        """Removes the given game from users greylist.
-        If game is None, all games will be removed.
-        If user is not on list, None will be returned.
-        If user was completely removed, True will be returned, otherwise False.
-        """
-        if game is None:
-            game = enums.GreylistGames.ALL
-        was_removed = None
-        if self.is_user_on_greylist(user, game):
-            self.gl_conf()[user.id] = self.gl_conf()[user.id] & ~game
-            was_removed = False
-            if self.gl_conf()[user.id] is enums.GreylistGames.No_Game:
-                del (self.gl_conf()[user.id])
-                was_removed = True
-        Config().save(self.plugin)
-        return was_removed
-
-    def get_greylist_names(self):
-        """Returns the greylisted members"""
-        greylisted_members = ", ".join([self.plugin.bot.get_user(uid).name for uid in self.gl_conf()])
-        return greylisted_members
-
-    def is_user_on_greylist(self, user: discord.Member, game: enums.GreylistGames):
-        """Returns if user is for the given game on the greylist"""
-        return self.is_userid_on_greylist(user.id, game)
-
-    def is_userid_on_greylist(self, userid: int, game: enums.GreylistGames):
-        """Returns if user id is for the given game on the greylist"""
-        if userid in self.gl_conf():
-            if self.gl_conf()[userid] & game is not 0:
-                return True
-        return False
-
-
-class CommandDisable:
-    """Manage disabling commands in certain channels
-    The cmds and channels will be saved as tuple:
-    [0]: command [1]: channel id [2]: expiring time
-    """
-
-    def __init__(self, plugin):
-        self.plugin = plugin
-
-    def cd_conf(self):
-        return Config().get(self.plugin)['disabled_cmds']
-
-    def check_expired(self):
-        """Checks all disabled commands and removes expired disablings"""
-        to_remove = []
-        for cmd_tuple in self.cd_conf():
-            if cmd_tuple[2] < datetime.datetime.now():
-                to_remove.append(cmd_tuple)
-
-        for removing in to_remove:
-            self.cd_conf().remove(removing)
-
-        Config().save(self.plugin)
-
-    def disable(self, command, channel: discord.TextChannel, hours: int = 0):
-        """Disables the given command in the given channel for given hours.
-        If cmd in channel was now disabled, True will be returned,
-        if cmd already disabled in channel False.
-        """
-        self.check_expired()
-
-        if command.startswith("!"):
-            command = command[1:]
-        if hours < 1:
-            exp_time = datetime.datetime.max
-        else:
-            exp_time = datetime.datetime.now() + datetime.timedelta(hours=hours)
-
-        is_adding = True
-        for t in self.cd_conf():
-            if t[0] == command and t[1] == channel.id:
-                is_adding = False
-
-        if is_adding:
-            self.cd_conf().append((command, channel.id, exp_time))
-
-        Config().save(self.plugin)
-        return is_adding
-
-    def enable(self, command, channel: discord.TextChannel):
-        """Enables the given command in given channel.
-        If cmd is now enabled in given channel, True will be returned, otherwise False.
-        """
-        self.check_expired()
-
-        if command.startswith("!"):
-            command = command[1:]
-
-        to_remove = None
-        for t in self.cd_conf():
-            if t[0] == command and t[1] == channel.id:
-                to_remove = t
-
-        is_removing = False
-        if to_remove is not None:
-            self.cd_conf().remove(to_remove)
-            is_removing = True
-
-        Config().save(self.plugin)
-        return is_removing
-
-    def can_cmd_executed(self, command, channel: discord.TextChannel):
-        """Returns if command can be executed in given channel"""
-        return self.can_cmd_executed_id(command, channel.id)
-
-    def can_cmd_executed_id(self, command, channel_id: int):
-        """Returns if command can be executed in channel with given id"""
-        self.check_expired()
-        current_time = datetime.datetime.now()
-
-        for tp in self.cd_conf():
-            if tp[0] == command and tp[1] == channel_id and tp[2] > current_time:
-                return False
-        return True
+from subsystems.blacklist import Blacklist
+from subsystems.greylist import Greylist
+from subsystems.cmddisable import CommandDisable
 
 
 class Plugin(BasePlugin, name="Bot Management Commands"):
